Improved performance of wave calculations
Replaced float-based calculation by integer calculation.
The impact is signficant. Before this, Level 1 at 20 FPS resulted in 80% CPU load. Now 25 FPS is possible, resulting in similar load.

+++ BumbleBots/Utils.h
@@ -90,7 +90,7 @@ extern const int8_t rowDelta[];
 float smoothClamp(float value);
 float smoothStep(float value);
 
-float fastCos(float value);
+int16_t fastCos(int16_t value);
 
 inline int8_t sign(int8_t value) {
   return (value > 0) - (value < 0);
